+++ app/src/main/java/org/simple/clinic/teleconsultlog/teleconsultrecord/screen/TeleconsultRecordScreenKey.kt
@@ -3,20 +3,17 @@ package org.simple.clinic.teleconsultlog.teleconsultrecord.screen
 import android.os.Parcelable
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 import java.util.UUID
 
 @Parcelize
 data class TeleconsultRecordScreenKey(
     val patientUuid: UUID,
     val teleconsultRecordId: UUID
-) : FullScreenKey, Parcelable {
+) : ScreenKey(), Parcelable {
 
   @IgnoredOnParcel
   override val analyticsName: String = "Teleconsult Record Screen"
 
-  override fun layoutRes(): Int {
-    return R.layout.screen_teleconsult_record
-  }
+  override fun instantiateFragment() = TeleconsultRecordScreen()
 }